+++ lib/models/src/position.g.dart
@@ -10,6 +10,7 @@ _$PositionImpl _$$PositionImplFromJson(Map<String, dynamic> json) =>
     _$PositionImpl(
       id: (json['id'] as num?)?.toInt(),
       name: json['name'] as String?,
+      roleId: json['roleId'] as String?,
       createdDate: json['createdDate'] == null
           ? null
           : DateTime.parse(json['createdDate'] as String),
@@ -19,5 +20,6 @@ Map<String, dynamic> _$$PositionImplToJson(_$PositionImpl instance) =>
     <String, dynamic>{
       'id': instance.id,
       'name': instance.name,
+      'roleId': instance.roleId,
       'createdDate': instance.createdDate?.toIso8601String(),
     };